more GUI changes. Added GUIButton and GUIText classes
