+++ pesukarhu/ui/UIFactories.h
@@ -16,27 +16,11 @@ namespace pk
 {
     namespace ui
     {
-        // *User defined ui events
-        class OnClickEvent
-        {
-        public:
-            virtual void onClick(InputMouseButtonName button) = 0;
-            virtual ~OnClickEvent(){}
-        };
-
-
         class InputFieldOnSubmitEvent
         {
         public:
             virtual void onSubmit(std::string inputFieldText) = 0;
             virtual ~InputFieldOnSubmitEvent() {}
-        };
-
-        struct UIFactoryButton
-        {
-            entityID_t rootEntity = 0;
-            entityID_t imgEntity = 0;
-            entityID_t txtEntity = 0;
         };
 
         struct UIFactoryInputField
@@ -50,29 +34,6 @@ namespace pk
             entityID_t rootEntity;
             entityID_t checkedStatusImgEntity;
         };
-
-        entityID_t create_text(
-            const std::string& str, const Font& font,
-            ConstraintProperties constraintProperties,
-            vec3 color = vec3(1.0f, 1.0f, 1.0f),
-            bool bold = false
-        );
-
-
-        UIFactoryButton create_button(
-            std::string txt, const Font& font,
-            ConstraintProperties constraintProperties,
-            float width, float height,
-            OnClickEvent* onClick,
-            bool selectable = false,
-            vec3 color = { 0.1f, 0.1f, 0.1f },
-            vec3 textColor = { 1, 1, 1 },
-            vec3 textHighlightColor = { 1, 1, 1 },
-            vec3 backgroundHighlightColor = { 0.2f, 0.2f, 0.2f },
-            GUIFilterType filter = GUIFilterType::GUI_FILTER_TYPE_NONE,
-            Texture* pTexture = nullptr,
-            vec4 textureCropping = vec4(0, 0, 1, 1)
-        );
 
         UIFactoryInputField create_input_field(
             std::string infoTxt,
